+++ pages/business/handle.tsx
@@ -10,7 +10,7 @@ import InteractiveHandle from "../../components/interactive/InteractiveHandle";
 
 const Handle: FC = () => {
   return (
-    <div className="flex items-center justify-center w-full h-full">
+    <div className="flex items-center justify-center w-full h-screen">
       <InteractiveHandle />
     </div>
   );
