+++ pages/index.tsx
@@ -1,269 +1,204 @@
-import {
-  useMotionTemplate,
-  useScroll,
-  useSpring,
-  useTransform,
-} from "framer-motion";
-import type { FC } from "react";
-import { manifest } from "../common/Manifest";
-import Activity from "../components/Activity";
-import Hyperlink from "../components/Hyperlink";
-import Quote from "../components/Quote";
+//
+//  cinderbloc.tsx
+//  d-exclaimation
+//
+//  Created by d-exclaimation on 25 Jan 2023
+//
+
+import * as Progress from "@radix-ui/react-progress";
+import Link from "next/link";
+import { FC, useEffect, useState } from "react";
+import { manifest, Stories } from "../common/Manifest";
 import Scrambled from "../components/Scrambled";
-import Signature from "../components/Signature";
 import { withHead } from "../hoc/withHead";
+import { useScrollSelection } from "../hooks/minigames/cinderbloc/useScrollSelection";
+import { useSelectScrolling } from "../hooks/minigames/cinderbloc/useSelectScrolling";
+
+const StoryTitles = Object.keys(manifest.stories) as Array<keyof Stories>;
+const StoryElements = Object.entries(manifest.stories).map(
+  ([title, { img, ...other }]) => {
+    return {
+      title: title as keyof Stories,
+      img,
+      ...other,
+    };
+  }
+);
 
 const Home: FC = () => {
-  const { scrollYProgress } = useScroll();
+  const [loading, setLoading] = useState(true);
+  const [progress, setProgress] = useState(0);
 
-  // Motion for title
-  const titleTransformation = useTransform(scrollYProgress, [0, 0.4], [6, 3]);
-  const titleSpring = useSpring(titleTransformation, {
-    mass: 0.008,
-  });
-  const titleSize = useMotionTemplate`${titleSpring}vw`;
+  // Mobile scrolling
+  const scrolled = useScrollSelection(StoryTitles);
+
+  // Desktop select scrolling
+  const [{ mouse, selected }, { onHover, onMouseMove, onClear }] =
+    useSelectScrolling(StoryTitles);
+
+  useEffect(() => {
+    // Loading animation
+    setProgress(30);
+    const timeouts = [
+      setTimeout(() => setProgress(100), 500),
+      setTimeout(() => setLoading(false), 1000),
+    ];
+
+    // Preload images
+    StoryElements.forEach(({ img }) => {
+      if (!img) return;
+      const preloader = new Image();
+      preloader.src = img;
+    });
+
+    return () => timeouts.forEach((timeout) => clearTimeout(timeout));
+  }, []);
 
   return (
-    <div className="w-full h-full">
-      <div className="fixed inset-0">
-        <div className="max-w-[1000px] my-0 mx-auto py-[14px] px-[12px]">
-          <div className="flex flex-col items-start">
-            {/* Title */}
-            <Scrambled
-              phrases={["vin", "vincent", "d-exclaimation"]}
-              align="items-start"
-              justify="justify-center"
-              style={{ fontSize: titleSize }}
+    <>
+      {/* Loading  */}
+      {loading && (
+        <div className="flex items-center justify-center transition-transform w-screen h-screen">
+          <Progress.Root className="overflow-hidden rounded-full max-w-xl w-[75vw] h-6 [transform:translateZ(0)] bg-neutral-200">
+            <Progress.Indicator
+              className="bg-neutral-800 w-full h-full transition-transform duration-500"
+              style={{ transform: `translateX(-${100 - progress}%)` }}
             />
-            {/* Socials */}
-            <div className="flex flex-row md:flex-col items-start text-xs md:text-base flex-wrap">
-              {manifest.links.map((social) => (
-                <div key={social.href}>
-                  <Hyperlink {...social} />
-                </div>
-              ))}
-            </div>
-          </div>
+          </Progress.Root>
         </div>
-      </div>
-      <div
-        className="
-        relative 
-        my-0 mx-auto z-10 
-        p-4 md:p-16 
-        max-w-[90vw] md:max-w-[min(800px,95vw)] 
-        top-[12vh] md:top-[20vh] 
-        md:left-[8vw] 
-        text-sm md:text-lg 
-        bg-gray-50"
-      >
-        {/* Brief and description */}
-        <div>
-          <Scrambled
-            phrases={[
-              "Software engineer",
-              "Computer science student",
-              "Professional memer",
-              "Pineaple on pizza enjoyer",
-            ]}
-            speed={24}
-            delay={5_000}
-            align="items-start"
-            justify="justify-start"
-            className="font-bold text-lg md:text-xl mb-2"
-          />
-          {/* Brief */}
-          <Scrambled
-            phrases={[
-              "Brief",
-              "Software",
-              "Engineering",
-              "Career",
-              "Meaning",
-              "Goal",
-              "Brief",
-            ]}
-            speed={2}
-            delay={1_000}
-            align="items-start"
-            justify="justify-start"
-            className="font-semibold md:text-lg my-2"
-            color={{
-              dud: "text-slate-400",
-              normal: "text-indigo-500",
-            }}
-          />
-          <p className="text-sm md:text-base font-extralight font-sans">
-            <Quote>Software engineer improving his craft</Quote>
-            <br />
-            Short description of myself would be an enthusiastic and career
-            driven person who enjoy the art of software development and want to
-            use that to build something meaningful.
-          </p>
-          {/* Catching up */}
-          <Scrambled
-            phrases={[
-              "Progress",
-              "Progress",
-              "Catching up",
-              "Studies",
-              "Life",
-              "Progress",
-              "Progress",
-            ]}
-            speed={2}
-            delay={1_000}
-            align="items-start"
-            justify="justify-start"
-            className="font-semibold md:text-lg my-2"
-            color={{
-              dud: "text-slate-400",
-              normal: "text-violet-500",
-            }}
-          />
-          <p className="text-sm md:text-base font-extralight font-sans">
-            <Quote>Computer science student breaking into tech</Quote>
-            <br />A huge portion of my life at the moment is about my studies
-            and my future career. I am currently a university student studying
-            computer science. One that will be graduating by the end of 2023 and
-            fully dedicate my life toward my career.
-          </p>
-        </div>
-        {/* Work experiences */}
-        <div className="mt-5 mb-2">
-          <Scrambled
-            phrases={[
-              "Professional experience",
-              "Work history",
-              "Projects",
-              "Things I put on my resume",
-            ]}
-            speed={24}
-            delay={5_000}
-            align="items-start"
-            justify="justify-start"
-            className="font-bold text-lg md:text-xl mb-2"
-          />
-        </div>
-        {/* Work */}
-        <Scrambled
-          phrases={[
-            "Experiences",
-            "Work",
-            "Walk of life",
-            "Roles",
-            "Experiences",
-          ]}
-          speed={2}
-          delay={1_000}
-          align="items-start"
-          justify="justify-start"
-          className="font-semibold md:text-lg my-2"
-          color={{
-            dud: "text-slate-400",
-            normal: "text-blue-500",
-          }}
-        />
-        <Activity
-          subtitle="Software Engineer Intern"
-          title={{ name: "Partly", href: "https://partly.com" }}
-          time={{
-            __t: "specified",
-            start: "November 2022",
-            end: "February 2023",
-          }}
-        >
-          Worked on{" "}
-          <a
-            className="text-sky-500 decoration-cyan-500 active:underline hover:underline cursor-pointer"
-            href="https://partspal.partly.com"
-          >
-            Partspal
-          </a>{" "}
-          and marketplace integations microservices using Typescript, Node.js,
-          React, and PostgreSQL
-        </Activity>
-        <Activity
-          subtitle="Software Engineer Contract"
-          title={{ name: "Zentax", href: "https://zentax.id" }}
-          time={{
-            __t: "specified",
-            start: "February 2021",
-            end: "November 2021",
-          }}
-        >
-          Worked on building{" "}
-          <a
-            className="text-sky-500 decoration-cyan-500 hover:underline active:underline cursor-pointer"
-            href="https://att.zentax.id"
-          >
-            their online attendance recording system
-          </a>{" "}
-          using Typescript, Node.js, React, GraphQL, and PostgreSQL
-        </Activity>
-        {/* Projects */}
-        <Scrambled
-          phrases={["Projects", "Open source", "Creation", "Projects"]}
-          speed={2}
-          delay={1_000}
-          align="items-start"
-          justify="justify-start"
-          className="font-semibold md:text-lg my-2"
-          color={{
-            dud: "text-slate-400",
-            normal: "text-teal-500",
-          }}
-        />
-        <Activity
-          subtitle="GraphQL server for Swift"
-          title={{ name: "Pioneer", href: "https://pioneer.dexclaimation.com" }}
-          time={{
-            __t: "ongoing",
-            start: "November 2022",
-          }}
-          color={{
-            border: "border-l-teal-200",
-            decoration: "decoration-teal-400",
-            text: "text-teal-400",
-          }}
-        >
-          An open source, spec-compliant GraphQL server for Swift on the server
-        </Activity>
-        <Activity
-          subtitle="All about me in a page"
-          title={{
-            name: "dexclaimation.com",
-            href: "https://dexclaimation.com",
-          }}
-          time={{
-            __t: "ongoing",
-            start: "February 2021",
-          }}
-          color={{
-            border: "border-l-teal-200",
-            decoration: "decoration-teal-400",
-            text: "text-teal-400",
-          }}
-        >
-          This website
-        </Activity>
-        <span className="font-sans font-extralight text-xs md:text-sm">
-          and more on{" "}
-          <a
-            className="text-sky-500 decoration-cyan-500 hover:underline active:underline cursor-pointer"
-            href="https://resume.dexclaimation.com"
-          >
-            my cv
-          </a>
-        </span>
+      )}
 
-        <Signature className="flex justify-end items-center px-2" />
+      {/* Page */}
+      <div
+        className={`transition-opacity duration-700 ${
+          loading ? "scale-0 opacity-0" : "scale-100 opacity-100"
+        }`}
+      >
+        <div className="opacity-0 h-[25vh]" />
+
+        {/* Mobile View */}
+        <nav
+          id="mobile-nav"
+          className="flex md:hidden flex-col p-[clamp(1rem,5vw,6rem)] gap-10"
+        >
+          {StoryElements.map(({ title, img, ...props }, i) => (
+            <Link
+              className={`
+              font-sans text-4xl no-underline relative
+              text-center md:text-start
+              ${scrolled === title ? "text-blue-100" : "text-black/25"}
+              `}
+              key={i}
+              {...props}
+            >
+              <div
+                className={`border-y-2 z-[2] py-2 ${
+                  scrolled === title ? "border-y-black" : "border-y-transparent"
+                }`}
+              >
+                <Scrambled
+                  className={`block relative z-[2] ${
+                    scrolled === title ? "opacity-100" : "opacity-25"
+                  }`}
+                  color={{
+                    dud: scrolled === title ? "text-neutral-400" : "text-black",
+                    normal:
+                      scrolled === title
+                        ? img
+                          ? "text-white"
+                          : "text-black"
+                        : "text-black",
+                  }}
+                  align="items-start"
+                  justify="justify-center"
+                  phrases={[title, title]}
+                />
+              </div>
+              {img && (
+                <img
+                  className={`
+                  absolute -translate-y-1/2 
+                  [transition:_transform_250ms,_opacity_250ms] pointer-events-none w-[min(90vw,800px)]
+                  ${
+                    scrolled === title
+                      ? "opacity-100 scale-100"
+                      : "opacity-0 scale-50"
+                  }`}
+                  alt={title}
+                  src={img}
+                />
+              )}
+            </Link>
+          ))}
+        </nav>
+
+        {/* Desktop View */}
+        <nav
+          id="desktop-nav"
+          className="hidden md:flex flex-col gap-2 p-[clamp(1rem,5vw,6rem)]"
+          onMouseMove={onMouseMove}
+          onMouseLeave={onClear}
+        >
+          {StoryElements.map(({ title, img, ...props }, i) => (
+            <a
+              className={`
+            font-sans text-[clamp(3rem,6vw,8rem)] font-light no-underline relative
+            text-center md:text-start
+            ${selected === title ? "text-blue-100" : "text-black/25"}
+            `}
+              key={i}
+              onMouseMove={onHover(title)}
+              {...props}
+            >
+              <div
+                className={`border-y-2 z-[2] ${
+                  selected === title ? "border-y-black" : "border-y-transparent"
+                }`}
+              >
+                <Scrambled
+                  className={`block relative z-[2] ${
+                    selected === title ? "opacity-100" : "opacity-25"
+                  }`}
+                  color={{
+                    dud: selected === title ? "text-neutral-400" : "text-black",
+                    normal:
+                      selected === title
+                        ? img
+                          ? "text-white"
+                          : "text-black"
+                        : "text-black",
+                  }}
+                  align="items-start"
+                  justify="justify-start"
+                  phrases={[title, title]}
+                />
+              </div>
+              {img && (
+                <img
+                  className={`
+                  absolute -translate-y-1/2 -translate-x-1/2 z-0
+                  [transition:_transform_250ms,_opacity_250ms] pointer-events-none w-[min(90vw,800px)]
+                  ${
+                    selected === title
+                      ? "opacity-100 scale-100"
+                      : "opacity-0 scale-50"
+                  }
+                  `}
+                  alt={title}
+                  style={mouse}
+                  src={img}
+                />
+              )}
+            </a>
+          ))}
+        </nav>
+        <div className="h-[40vh] md:h-0" />
       </div>
-    </div>
+    </>
   );
 };
 
 export default withHead(Home, {
   title: "d-exclaimation",
-  description: "Quick porfolio",
+  description: "d-exclaimation",
 });
